--- api/src/models/gcnotify.ts
@@ -1,16 +1,0 @@
-export interface IgcNotifyPostReturn {
-  content: object;
-  id: string;
-  reference: string;
-  scheduled_for: string;
-  template: object;
-  uri: string;
-}
-
-export interface IgcNotifyGenericMessage {
-  subject: string;
-  header: string;
-  body1: string;
-  body2: string;
-  footer: string;
-}
