+++ app/src/interfaces/useAdminApi.interface.ts
@@ -28,6 +28,7 @@ export interface IGetAccessRequestsListResponse {
 }
 
 export interface IgcNotifyGenericMessage {
+  subject: string;
   header: string;
   body1: string;
   body2: string;
